Update unit test
- Init project as an edp project

- Fix test
