+++ .edpproj/.gitignore
@@ -1,0 +1,1 @@
+# Please keep this file
