+++ module.conf
@@ -1,0 +1,46 @@
+{
+    "baseUrl": "src",
+    "paths": {},
+    "packages": [
+        {
+            "name": "eoo",
+            "location": "../dep/eoo/0.1.2/src",
+            "main": "main"
+        },
+        {
+            "name": "etpl",
+            "location": "../dep/etpl/3.0.0/src",
+            "main": "main"
+        },
+        {
+            "name": "fc-core",
+            "location": "../dep/fc-core/0.0.1-alpha.7/src",
+            "main": "main"
+        },
+        {
+            "name": "fc-storage",
+            "location": "../dep/fc-storage/0.0.1-alpha.6/src",
+            "main": "main"
+        },
+        {
+            "name": "mini-event",
+            "location": "../dep/mini-event/1.0.2/src",
+            "main": "main"
+        },
+        {
+            "name": "promise",
+            "location": "../dep/promise/1.0.2/src",
+            "main": "main"
+        },
+        {
+            "name": "underscore",
+            "location": "../dep/underscore/1.6.0/src",
+            "main": "underscore"
+        },
+        {
+            "name": "usertiming",
+            "location": "../dep/usertiming/0.1.6/dist",
+            "main": "usertiming.min"
+        }
+    ]
+}
