+++ test/context.html
@@ -1,0 +1,29 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset="utf-8">
+        <title>edp test runner - using jasmine</title>
+        <!--%css%-->
+        <!--%frameworks%-->
+        <script>
+            require.config(/*requireConfig*/);
+        </script>
+    </head>
+    <body>
+        <script>
+            (function () {
+                window.DEBUG = true;
+                require([ /*specs*/ ], function () {
+                    var runner = (parent || opener);
+                    runner.execute = function (socket) {
+                        var jasmineEnv = jasmine.getEnv();
+                        jasmineEnv.updateInterval = 1000;
+                        jasmineEnv.addReporter(new jasmine.EdpReporter(socket, runner));
+                        jasmineEnv.execute();
+                    };
+                    runner.start();
+                });
+            })();
+        </script>
+    </body>
+</html>
