+++ test/debug.html
@@ -1,0 +1,36 @@
+<!doctype html>
+<html>
+    <head>
+        <meta charset="utf-8">
+        <title>edp test runner - using jasmine</title>
+        <!--%css%-->
+        <!--%frameworks%-->
+        <script>
+            require.config(/*requireConfig*/);
+        </script>
+    </head>
+    <body>
+        <script>
+            (function () {
+                window.DEBUG = true;
+                require([ /*specs*/ ], function () {
+                    var jasmineEnv = jasmine.getEnv();
+                    jasmineEnv.updateInterval = 1000;
+                    var htmlReporter = jasmine.htmlReporter || new jasmine.HtmlReporter();
+                    jasmineEnv.addReporter(htmlReporter);
+                    // for jasmine2
+                    if (jasmine.htmlReporter) {
+                        htmlReporter.initialize();
+                    }
+                    // jasmine1.3.1
+                    else {
+                        jasmineEnv.specFilter = function(spec) {
+                          return htmlReporter.specFilter(spec);
+                        };
+                    }
+                    jasmineEnv.execute();
+                });
+            })();
+        </script>
+    </body>
+</html>
